Extract shared chart chrome into ChartStyle (Phase A.0, step 1: no-op dedup)
Create PerformanceMonitor.Ui/ChartStyle.cs as the single source of truth for
ScottPlot chart CHROME (figure/data backgrounds, axis/grid/legend colors, tick
label colors + font sizes).

ApplyThemeToChart, ReapplyAxisColors, and HandleChartFirstLoaded were byte-identical
between Dashboard TabHelpers and Lite ServerTab.Charts.cs (the "Matches Dashboard
... exactly" hand-maintained duplicate). Both now forward to ChartStyle. Pure no-op
dedup — zero visual change; both apps build green.

Color identity (the ChartColors palette / per-series colors) is intentionally NOT
moved — that is Phase A.3. The drifted theming copies (Lite history windows,
CorrelatedTimelineLanesControl, SetChartYLimitsWithLegendPadding) reconcile next via
a parameterized ChartStyle (showLegend option), screenshot-gated for their deltas.
