+++ Dashboard/Helpers/TabHelpers.cs
@@ -148,98 +148,16 @@ namespace PerformanceMonitorDashboard.Helpers
         }
 
         /// <summary>
-        /// Applies the current color theme to a ScottPlot chart.
+        /// Applies the current color theme (chrome) to a ScottPlot chart.
+        /// Delegates to the shared <see cref="ChartStyle"/> — single source of truth across apps.
         /// </summary>
-        public static void ApplyThemeToChart(WpfPlot chart)
-        {
-            ScottPlot.Color figureBackground, dataBackground, textColor, gridColor, legendBg, legendFg, legendOutline;
-
-            if (ThemeManager.CurrentTheme == "CoolBreeze")
-            {
-                figureBackground = ScottPlot.Color.FromHex("#EEF4FA");
-                dataBackground   = ScottPlot.Color.FromHex("#DAE6F0");
-                textColor        = ScottPlot.Color.FromHex("#1A2A3A");
-                gridColor        = ScottPlot.Color.FromHex("#A8BDD0").WithAlpha(120);
-                legendBg         = ScottPlot.Color.FromHex("#EEF4FA");
-                legendFg         = ScottPlot.Color.FromHex("#1A2A3A");
-                legendOutline    = ScottPlot.Color.FromHex("#A8BDD0");
-            }
-            else if (ThemeManager.HasLightBackground)
-            {
-                figureBackground = ScottPlot.Color.FromHex("#FFFFFF");
-                dataBackground   = ScottPlot.Color.FromHex("#F5F7FA");
-                textColor        = ScottPlot.Color.FromHex("#1A1D23");
-                gridColor        = ScottPlot.Colors.Black.WithAlpha(20);
-                legendBg         = ScottPlot.Color.FromHex("#FFFFFF");
-                legendFg         = ScottPlot.Color.FromHex("#1A1D23");
-                legendOutline    = ScottPlot.Color.FromHex("#DEE2E6");
-            }
-            else
-            {
-                figureBackground = ScottPlot.Color.FromHex("#22252b");
-                dataBackground   = ScottPlot.Color.FromHex("#111217");
-                textColor        = ScottPlot.Color.FromHex("#E4E6EB");
-                gridColor        = ScottPlot.Colors.White.WithAlpha(40);
-                legendBg         = ScottPlot.Color.FromHex("#22252b");
-                legendFg         = ScottPlot.Color.FromHex("#E4E6EB");
-                legendOutline    = ScottPlot.Color.FromHex("#2a2d35");
-            }
-
-            chart.Plot.FigureBackground.Color = figureBackground;
-            chart.Plot.DataBackground.Color = dataBackground;
-            chart.Plot.Axes.Color(textColor);
-            chart.Plot.Grid.MajorLineColor = gridColor;
-            chart.Plot.Legend.BackgroundColor = legendBg;
-            chart.Plot.Legend.FontColor = legendFg;
-            chart.Plot.Legend.OutlineColor = legendOutline;
-            chart.Plot.Legend.Alignment = ScottPlot.Alignment.LowerCenter;
-            chart.Plot.Legend.Orientation = ScottPlot.Orientation.Horizontal;
-            chart.Plot.Axes.Margins(bottom: 0); // No bottom margin - SetChartYLimitsWithLegendPadding handles Y-axis
-
-            // Explicitly set axis tick label colors (needed after DateTimeTicksBottom() is called)
-            chart.Plot.Axes.Bottom.TickLabelStyle.ForeColor = textColor;
-            chart.Plot.Axes.Left.TickLabelStyle.ForeColor = textColor;
-            chart.Plot.Axes.Bottom.Label.ForeColor = textColor;
-            chart.Plot.Axes.Left.Label.ForeColor = textColor;
-            chart.Plot.Axes.Bottom.TickLabelStyle.FontSize = 13;
-            chart.Plot.Axes.Left.TickLabelStyle.FontSize = 13;
-
-            // Set the WPF control Background to match so no white flash appears before ScottPlot's render loop fires
-            chart.Background = new SolidColorBrush(Color.FromRgb(figureBackground.R, figureBackground.G, figureBackground.B));
-
-            // Ensure ScottPlot renders with the correct colors the very first time it gets pixel dimensions.
-            // Without this, ScottPlot's first auto-render (triggered by SizeChanged) would show a white canvas
-            // before our FigureBackground color takes visual effect.
-            chart.Loaded -= HandleChartFirstLoaded;
-            if (!chart.IsLoaded)
-                chart.Loaded += HandleChartFirstLoaded;
-        }
-
-        private static void HandleChartFirstLoaded(object sender, RoutedEventArgs e)
-        {
-            var chart = (WpfPlot)sender;
-            chart.Loaded -= HandleChartFirstLoaded;
-            chart.Refresh();
-        }
+        public static void ApplyThemeToChart(WpfPlot chart) => ChartStyle.ApplyThemeToChart(chart);
 
         /// <summary>
-        /// Reapplies theme-appropriate text colors to chart axes.
-        /// Call this AFTER DateTimeTicksBottom() or other axis modifications.
+        /// Reapplies theme-appropriate text colors to chart axes (after DateTimeTicksBottom() etc.).
+        /// Delegates to the shared <see cref="ChartStyle"/>.
         /// </summary>
-        public static void ReapplyAxisColors(WpfPlot chart)
-        {
-            var textColor = ThemeManager.CurrentTheme == "CoolBreeze"
-                ? ScottPlot.Color.FromHex("#1A2A3A")
-                : ThemeManager.HasLightBackground
-                    ? ScottPlot.Color.FromHex("#1A1D23")
-                    : ScottPlot.Color.FromHex("#E4E6EB");
-            chart.Plot.Axes.Bottom.TickLabelStyle.ForeColor = textColor;
-            chart.Plot.Axes.Left.TickLabelStyle.ForeColor = textColor;
-            chart.Plot.Axes.Bottom.Label.ForeColor = textColor;
-            chart.Plot.Axes.Left.Label.ForeColor = textColor;
-            chart.Plot.Axes.Bottom.TickLabelStyle.FontSize = 13;
-            chart.Plot.Axes.Left.TickLabelStyle.FontSize = 13;
-        }
+        public static void ReapplyAxisColors(WpfPlot chart) => ChartStyle.ReapplyAxisColors(chart);
 
         /// <summary>
         /// Recursively finds all WpfPlot chart controls in a visual tree.
