+++ Lite/Controls/ServerTab.Charts.cs
@@ -1296,77 +1296,10 @@ public partial class ServerTab : UserControl
     }
 
     /// <summary>
-    /// Applies the Darling Data dark theme to a ScottPlot chart.
-    /// Matches Dashboard TabHelpers.ApplyThemeToChart exactly.
+    /// Applies the chrome theme to a ScottPlot chart.
+    /// Delegates to the shared <see cref="ChartStyle"/> — single source of truth across apps.
     /// </summary>
-    private static void ApplyTheme(ScottPlot.WPF.WpfPlot chart)
-    {
-        ScottPlot.Color figureBackground, dataBackground, textColor, gridColor, legendBg, legendFg, legendOutline;
-
-        if (ThemeManager.CurrentTheme == "CoolBreeze")
-        {
-            figureBackground = ScottPlot.Color.FromHex("#EEF4FA");
-            dataBackground   = ScottPlot.Color.FromHex("#DAE6F0");
-            textColor        = ScottPlot.Color.FromHex("#1A2A3A");
-            gridColor        = ScottPlot.Color.FromHex("#A8BDD0").WithAlpha(120);
-            legendBg         = ScottPlot.Color.FromHex("#EEF4FA");
-            legendFg         = ScottPlot.Color.FromHex("#1A2A3A");
-            legendOutline    = ScottPlot.Color.FromHex("#A8BDD0");
-        }
-        else if (ThemeManager.HasLightBackground)
-        {
-            figureBackground = ScottPlot.Color.FromHex("#FFFFFF");
-            dataBackground   = ScottPlot.Color.FromHex("#F5F7FA");
-            textColor        = ScottPlot.Color.FromHex("#1A1D23");
-            gridColor        = ScottPlot.Colors.Black.WithAlpha(20);
-            legendBg         = ScottPlot.Color.FromHex("#FFFFFF");
-            legendFg         = ScottPlot.Color.FromHex("#1A1D23");
-            legendOutline    = ScottPlot.Color.FromHex("#DEE2E6");
-        }
-        else
-        {
-            figureBackground = ScottPlot.Color.FromHex("#22252b");
-            dataBackground   = ScottPlot.Color.FromHex("#111217");
-            textColor        = ScottPlot.Color.FromHex("#E4E6EB");
-            gridColor        = ScottPlot.Colors.White.WithAlpha(40);
-            legendBg         = ScottPlot.Color.FromHex("#22252b");
-            legendFg         = ScottPlot.Color.FromHex("#E4E6EB");
-            legendOutline    = ScottPlot.Color.FromHex("#2a2d35");
-        }
-
-        chart.Plot.FigureBackground.Color = figureBackground;
-        chart.Plot.DataBackground.Color = dataBackground;
-        chart.Plot.Axes.Color(textColor);
-        chart.Plot.Grid.MajorLineColor = gridColor;
-        chart.Plot.Legend.BackgroundColor = legendBg;
-        chart.Plot.Legend.FontColor = legendFg;
-        chart.Plot.Legend.OutlineColor = legendOutline;
-        chart.Plot.Legend.Alignment = ScottPlot.Alignment.LowerCenter;
-        chart.Plot.Legend.Orientation = ScottPlot.Orientation.Horizontal;
-        chart.Plot.Axes.Margins(bottom: 0); /* No bottom margin - SetChartYLimitsWithLegendPadding handles Y-axis */
-
-        chart.Plot.Axes.Bottom.TickLabelStyle.ForeColor = textColor;
-        chart.Plot.Axes.Left.TickLabelStyle.ForeColor = textColor;
-        chart.Plot.Axes.Bottom.Label.ForeColor = textColor;
-        chart.Plot.Axes.Left.Label.ForeColor = textColor;
-        chart.Plot.Axes.Bottom.TickLabelStyle.FontSize = 13;
-        chart.Plot.Axes.Left.TickLabelStyle.FontSize = 13;
-
-        // Set the WPF control Background to match so no white flash appears before ScottPlot's render loop fires
-        chart.Background = new System.Windows.Media.SolidColorBrush(System.Windows.Media.Color.FromRgb(figureBackground.R, figureBackground.G, figureBackground.B));
-
-        // Ensure ScottPlot renders with the correct colors the very first time it gets pixel dimensions.
-        chart.Loaded -= HandleChartFirstLoaded;
-        if (!chart.IsLoaded)
-            chart.Loaded += HandleChartFirstLoaded;
-    }
-
-    private static void HandleChartFirstLoaded(object sender, RoutedEventArgs e)
-    {
-        var chart = (ScottPlot.WPF.WpfPlot)sender;
-        chart.Loaded -= HandleChartFirstLoaded;
-        chart.Refresh();
-    }
+    private static void ApplyTheme(ScottPlot.WPF.WpfPlot chart) => ChartStyle.ApplyThemeToChart(chart);
 
     private void OnThemeChanged(string _)
     {
@@ -1395,22 +1328,10 @@ public partial class ServerTab : UserControl
     }
 
     /// <summary>
-    /// Reapplies theme-appropriate text colors and font sizes after DateTimeTicksBottom() resets them.
+    /// Reapplies theme-appropriate axis text colors/sizes after DateTimeTicksBottom() resets them.
+    /// Delegates to the shared <see cref="ChartStyle"/>.
     /// </summary>
-    private static void ReapplyAxisColors(ScottPlot.WPF.WpfPlot chart)
-    {
-        var textColor = ThemeManager.CurrentTheme == "CoolBreeze"
-            ? ScottPlot.Color.FromHex("#1A2A3A")
-            : ThemeManager.HasLightBackground
-                ? ScottPlot.Color.FromHex("#1A1D23")
-                : ScottPlot.Color.FromHex("#E4E6EB");
-        chart.Plot.Axes.Bottom.TickLabelStyle.ForeColor = textColor;
-        chart.Plot.Axes.Left.TickLabelStyle.ForeColor = textColor;
-        chart.Plot.Axes.Bottom.Label.ForeColor = textColor;
-        chart.Plot.Axes.Left.Label.ForeColor = textColor;
-        chart.Plot.Axes.Bottom.TickLabelStyle.FontSize = 13;
-        chart.Plot.Axes.Left.TickLabelStyle.FontSize = 13;
-    }
+    private static void ReapplyAxisColors(ScottPlot.WPF.WpfPlot chart) => ChartStyle.ReapplyAxisColors(chart);
 
     /// <summary>
     /// Sets Y-axis limits with padding for bottom legend and top breathing room.
